pretify run_environment test
